requirements.txt
wheel
pyttsx3
SpeechRecognition
pywhatkit
wikipedia
requests
python-decouple